requirements.txt
streamlit==1.31.0
pydantic==2.6.1
google-generativeai==0.4.0
python-dotenv==1.0.1